pyproject.toml
[tool.poetry]
name = "inventory-management-system"
version = "0.1.0a"
description = "A simple inventory management system for a cyberpunk-themed game. The system will manage items that players can acquire in the game. These items could range from cybernetic enhancements to weapons and gadgets."
authors = ["Serghei Dabula <dabula.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
fastapi = "0.112.0"
uvicorn = "0.30.5"
alembic = "1.13.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
psycopg2-binary = "2.9.9"
asyncpg = "0.29.0"
fastapi-users = {extras = ["redis", "sqlalchemy"], version = "13.0.0"}

[tool.poetry.dev-dependencies]
pytest-asyncio = "0.23.8"
sqlalchemy-utils = "0.41.2"
httpx = "0.27.0"
pytest-mock = "3.14.0"
pytest-cov = "5.0.0"
python-dotenv = "1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
